Add a route to the local proxy for POST requests
There is now an endpoint to get data with a POST rather than a GET
with query parameters. This commit allows us to use this endpoint
while developing locally.

I am having trouble getting the POST endpoint to work correctly and
consistently, so I don't want to merge this yet until that is
resolved.