refactor(tabs-container): remove tabs-container listener to global filter change
I want tabs-container to be the component in charge of filters, but due to some complications with
ui-router I have to pass data through a factory. Decision was made to have tabs-container assume it
doesn't have to listen for filter changes, because it should be the only one making filter changes.
Of course, this could be changed pretty easily.